+++ apps/docs/sidebars.ts
@@ -205,6 +205,11 @@ const sidebars: SidebarsConfig = {
             },
             {
               type: 'doc',
+              id: 'components/inputs/SegmentedControl/segmentedControl',
+              label: 'SegmentedControl (Deprecated)',
+            },
+            {
+              type: 'doc',
               id: 'components/inputs/SelectChipAlpha/selectChipAlpha',
               label: 'SelectChip (Alpha)',
             },